Project/postagram-backend/requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==21.1



